backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
easyocr
opencv-python
numpy
